requirements.txt
streamlit
pandas
openpyxl
pdfplumber
python-docx
docxtpl
matplotlib
plotly
pyright
